Change queries for removing additional rows which have concept_id=0

+++ inst/sql/sql server/000.OMOP CDM sql server ddl.sql
@@ -370,7 +370,7 @@ CREATE TABLE procedure_occurrence
 CREATE TABLE drug_exposure
 (
   drug_exposure_id				      bigint			  NOT NULL ,
-  person_id						          INTEGER			  NOT NULL ,
+  person_id						          bigint			  NOT NULL ,
   drug_concept_id				        INTEGER			  NOT NULL ,
   drug_exposure_start_date		  DATE			    NOT NULL ,
   drug_exposure_start_datetime  DATETIME2		  NULL ,
